# Fix min_pos when all negative + speed up