--- a/sklearn/utils/arrayfuncs.pyx
+++ b/sklearn/utils/arrayfuncs.pyx
@@ -18,36 +18,27 @@ ctypedef np.float64_t DOUBLE
 np.import_array()
 
 
-def min_pos(np.ndarray X):
-   """
-   Find the minimum value of an array over positive values
-
-   Returns a huge value if none of the values are positive
-   """
-   if X.dtype.name == 'float32':
-      return _float_min_pos(<float *> X.data, X.size)
-   elif X.dtype.name == 'float64':
-      return _double_min_pos(<double *> X.data, X.size)
-   else:
-      raise ValueError('Unsupported dtype for array X')
-
-
-cdef float _float_min_pos(float *X, Py_ssize_t size):
-   cdef Py_ssize_t i
-   cdef float min_val = DBL_MAX
-   for i in range(size):
-      if 0. < X[i] < min_val:
-         min_val = X[i]
-   return min_val
+def min_pos(np.ndarray X): 
+   """ 
+   Find the minimum value of an array over positive values 
+ 
+   Returns a huge value if none of the values are positive 
+   """ 
+   if X.dtype == np.float32: 
+      return _min_pos[float](<float *> X.data, X.size) 
+   elif X.dtype == np.float64: 
+      return _min_pos[double](<double *> X.data, X.size) 
+   else: 
+      raise ValueError('Unsupported dtype for array X') 
 
 
-cdef double _double_min_pos(double *X, Py_ssize_t size):
-   cdef Py_ssize_t i
-   cdef np.float64_t min_val = FLT_MAX
-   for i in range(size):
-      if 0. < X[i] < min_val:
-         min_val = X[i]
-   return min_val
+cdef floating _min_pos(floating* X, Py_ssize_t size): 
+   cdef Py_ssize_t i 
+   cdef floating min_val = FLT_MAX if floating is float else DBL_MAX 
+   for i in range(size): 
+      if 0. < X[i] < min_val: 
+         min_val = X[i] 
+   return min_val 
 
 
 # General Cholesky Delete.